Измененна структура базы данных

+++ src/main/java/com/example/wms/wms/repositories/ProductRepository.java
@@ -13,7 +13,7 @@ public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
     @Transactional
     @Query("update ProductEntity p set p.count_on_shipping = ?2 , " +
             "p.count_on_warehouse = ?3 where p.id = ?1")
-    void updateById(Long id, int count_on_shipping, int count_on_warehouse);
+    void updateById(Long id, double count_on_shipping, double count_on_warehouse);
 
     @Query("select p from ProductEntity p where p.product_name = ?1")
     List<ProductEntity> getProductByName(String name);
